Commit 1
1) Moved styles from DatePickerFormUser.styled.jsx to FormUser.styled.jsx;
2) Deleted DatePickerFormUser.styled.jsx;
3) Updated FormUser.js and DatePickerFormUser.js for letting update request from form go;
