--- src/components/Form/FormUser/DatePickerFormUser/DatePickerFormUser.styled.jsx
@@ -1,162 +1,0 @@
-import styled from 'styled-components';
-
-const viewport = {
-  mob: '@media screen and (max-width: 767px)',
-};
-
-
-export const Label = styled.label`
-  display: flex;
-  flex-direction: column;
-  margin-bottom: 0.5rem;
-
-  font-family: 'InterRegular';
-  font-size: 14px;
-  line-height: 15px;
-
-  color: ${props => props.theme.variable.bgReversLabel};
-`;
-
-export const DatePickerFormUserWrapper = styled.div`
-  font-family: 'Inter', sans-serif;
-  margin-right: 8px;
-  & .react-datepicker__today-button {
-    border: none;
-    background: none;
-  }
-  & input {
-    display: inline;
-
-    font-family: 'InterSemiBold';
-    font-size: 16px;
-    background-color: ${props => props.theme.variable.secondaryBg};
-    outline: none;
-    border: 1px solid ;
-    border-color: ${props => props.theme.variable.borderProfileColor};
-    border-radius: 8px;
-
-
-    text-align: center;
-    padding: 0;
-
-
-    height: 42px;
-
-    width: 299px;
-
-    color: white;
-    text-transform: uppercase;
-
-    font-weight: 700;
-    line-height: calc(14 / 18);
-
-    ${viewport.mob} {
-      height: 46px;
-  
-      width: 354px;
-    
-    }
-  }
-
-  & .react-datepicker {
-    background-color: #3e85f3;
-    color: white;
-    border-radius: 16px;
-  }
-  & .react-datepicker__header {
-    background-color: #3e85f3;
-    border-top-left-radius: 16px;
-    border-top-right-radius: 16px;
-    border-color: rgba(255, 255, 255, 0.5);
-  }
-  & .react-datepicker__triangle::after,
-  & .react-datepicker__triangle::before {
-    background-color: #3e85f3;
-    display: none;
-  }
-  & .react-datepicker__current-month {
-    font-size: 16px;
-    margin-bottom: 18px;
-  }
-  & .react-datepicker__day-name,
-  & .react-datepicker__day.react-datepicker__day--outside-month {
-    color: rgba(255, 255, 255, 0.5);
-  }
-
-  & .react-datepicker__navigation-icon::before {
-    border-color: white;
-  }
-  & .react-datepicker__current-month {
-    color: white;
-  }
-  & .react-datepicker__week:last-of-type {
-    margin-bottom: 0;
-  }
-  & .react-datepicker__week:not(:last-of-type) {
-    margin-bottom: 7px;
-  }
-  & .react-datepicker__day.react-datepicker__day--selected {
-    border-radius: 50%;
-    color: #3e85f3;
-
-    background-color: rgba(255, 255, 255, 0.5);
-  }
-  & .react-datepicker__day.react-datepicker__day--today {
-    background-color: #fff;
-    border-radius: 50%;
-    color: #3e85f3;
-    font-weight: 500;
-  }
-  & .react-datepicker__day--keyboard-selected {
-    border: none;
-    background: none;
-  }
-  & .react-datepicker__day {
-    color: inherit;
-    font-size: 14px;
-  }
-
-`;
-
-
-
-export const ControlWrapper = styled.div`
-  display: flex;
-  justify-content: space-between;
-  margin-bottom: 24px;
-
-  ${viewport.mob} {
-    margin-bottom: 18px;
-  }
-`;
-
-export const Controls = styled.button`
-display:none;
-  display: inline-flex;
-  justify-content: center;
-  align-items: center;
-  width: 36px;
-  height: 30px;
-  background-color: ${props => props.theme.variable.CalendarLinkColor};
-  border: ${props => props.theme.variable.borderColorWeekCalendar};
-  color: ${props => props.theme.variable.calendarTextColor};
-
-  &:disabled {
-    background-color: ${({ theme }) => theme.variable.mainBackgroundColor};
-    color: ${({ theme }) => theme.variable.activeArrowColor};
-  }
-
-  @media (min-width: 768px) {
-    width: 38px;
-  }
-  &:first-of-type {
-    border-top-left-radius: 8px;
-    border-bottom-left-radius: 8px;
-    border-right-width: 0.5px;
-  }
-  &:last-of-type {
-    border-top-right-radius: 8px;
-    border-bottom-right-radius: 8px;
-    border-left-width: 0.5px;
-  }
-`;
